pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.1.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
